feat: add migrations with changing models and configurations

+++ SportAPI/Sport/Models/Match.cs
@@ -7,12 +7,12 @@ namespace SportAPI.Sport.Models
 {
   public class Match : DomainEntity
   {
-    public int Id { get; set; }
+    public long Id { get; set; }
     public string TeamOne { get; set; }
     public string TeamTwo { get; set; }
     public bool InHouse { get; set; }
     public DateTime DateOfMatch { get; set; }
-    public int SportClubId { get; set; }
+    public long SportClubId { get; set; }
     public virtual SportClub SportClub { get; set; }
   }
 }
