+++ SportAPI/Sport/Models/Training.cs
@@ -7,11 +7,11 @@ namespace SportAPI.Sport.Models
 {
   public class Training : DomainEntity
   {
-    public int Id { get; set; }
+    public long Id { get; set; }
     public string Name { get; set; }
     public string Description { get; set; }
     public DateTime TimeOfTraining { get; set; }
-    public int SportClubId { get; set; }
+    public long SportClubId { get; set; }
     public virtual SportClub SportClub { get; set; }
   }
 }
